(feat) Review fixes
- Added `RowGroupsMerger` for efficient k-way sort-merge of sorted Parquet inputs.
- Introduced new merge logic with clustering, heap-based row selection, and stable tie-breaking order.
- Updated Helm notes to clarify `jobsmanager.scanIntervalSecs` parameter for compaction service.
