+++ config/helm/icegate/templates/NOTES.txt
@@ -42,6 +42,6 @@ OTLP HTTP (port {{ .Values.ingest.otlpHttp.port }}):
 === Compaction Service ===
 
 Long-running Parquet compaction (rewrites small data files into larger ones,
-scanning every {{ .Values.compact.compaction.scanIntervalSecs }}s). It has no API; follow its progress with:
+scanning every {{ .Values.compact.compaction.jobsmanager.scanIntervalSecs }}s). It has no API; follow its progress with:
   kubectl logs -n {{ .Release.Namespace }} -l app.kubernetes.io/component=compact -f
 {{- end }}
